requirements.txt
python-telegram-bot==21.3
solana==0.34.2
requests==2.32.2